param is a 16 bit integer. implement proper llc encapsulation/decapsulation

+++ src/devices/csma-cd/csma-cd-net-device.h
@@ -222,7 +222,7 @@ protected:
    * \return Returns true if the packet should be forwarded up the
    * protocol stack.
    */
-  bool ProcessHeader (Packet& p, int& param);
+  bool ProcessHeader (Packet& p, uint16_t & param);
 
 private:
   // disable copy constructor and operator =
